+++ deepinv/training/trainer.py
@@ -22,7 +22,9 @@ class Trainer:
     r"""Trainer(model, physics, optimizer, train_dataloader, ...)
     Trainer class for training a reconstruction network.
 
-    See the :ref:`User Guide <trainer>` for more details on how to adapt the trainer to your needs.
+    .. seealso::
+
+        See the :ref:`User Guide <trainer>` for more details and for how to adapt the trainer to your needs.
 
     Training can be done by calling the :func:`deepinv.Trainer.train` method, whereas
     testing can be done by calling the :func:`deepinv.Trainer.test` method.
@@ -37,8 +39,9 @@ class Trainer:
     dictionary of the model, ``loss`` the loss history, ``optimizer`` the state dictionary of the optimizer,
     and ``eval_metrics`` the evaluation metrics history.
 
-    - Use :func:`deepinv.Trainer.get_samples_online` when measurements are simulated from a ground truth returned by the dataloader.
-    - Use :func:`deepinv.Trainer.get_samples_offline` when both the ground truth and measurements are returned by the dataloader (and also optionally physics generator params).
+    - Use `online_measurements=True` when measurements are simulated online from a ground truth returned by the dataloader.
+    - Use `online_measurements=False` when both the ground truth and measurements are returned by the dataloader (and also optionally physics generator params).
+      This could be from a dataset generated using :class:`deepinv.datasets.generate_dataset`.
 
     .. note::
 
@@ -53,14 +56,17 @@ class Trainer:
         as long as it takes as input ``(x, x_net, y, physics, model)`` and returns a scalar, where ``x`` is the ground
         reconstruction, ``x_net`` is the network reconstruction :math:`\inversef{y}{A}`,
         ``y`` is the measurement vector, ``physics`` is the forward operator
-        and ``model`` is the reconstruction network. Note that not all inpus need to be used by the loss,
+        and ``model`` is the reconstruction network. Note that not all inputs need to be used by the loss,
         e.g., self-supervised losses will not make use of ``x``.
 
     .. warning::
 
-        If a physics generator is used to generate params for online measurements, the generated params will vary each epoch.
-        If this is not desired (you want the same online measurements each epoch), set ``loop_physics_generator=True``.
-        Caveat: this requires ``shuffle=False`` in your dataloaders.
+        If a physics generator or a noise model is used to generate random params for online measurements, the generated measurements will randomly vary each epoch.
+        If this is not desired (i.e. you want the same online measurements each epoch), set ``loop_random_online_physics=True``.
+        This resets the physics generator and noise model's random generators every epoch.
+
+        **Caveat**: this requires ``shuffle=False`` in your dataloaders.
+
         An alternative, safer solution is to generate and save params offline using :func:`deepinv.datasets.generate_dataset`.
         The params dict will then be automatically updated every time data is loaded.
 
@@ -70,13 +76,15 @@ class Trainer:
     :param int epochs: Number of training epochs. Default is 100.
     :param torch.optim.Optimizer optimizer: Torch optimizer for training the network.
     :param torch.utils.data.DataLoader, list[torch.utils.data.DataLoader] train_dataloader: Train data loader(s) should provide a
-        a signal x or a tuple of (x, y) signal/measurement pairs.
+        a signal `x` or a tuple of `(x, y)` signal/measurement pairs or a tuple `(x, y, params)`
+        where `params` is a dict of physics generator parameters to be loaded into the physics each iteration.
     :param deepinv.loss.Loss, list[deepinv.loss.Loss] losses: Loss or list of losses used for training the model.
         Optionally wrap losses using a loss scheduler for more advanced training.
         :ref:`See the libraries' training losses <loss>`. By default, it uses the supervised mean squared error.
         Where relevant, the underlying metric should have ``reduction=None`` as we perform the averaging using :class:`deepinv.utils.AverageMeter` to deal with uneven batch sizes.
     :param None, torch.utils.data.DataLoader, list[torch.utils.data.DataLoader] eval_dataloader: Evaluation data loader(s)
-        should provide a signal x or a tuple of (x, y) signal/measurement pairs.
+        should provide a signal x or a tuple of (x, y) signal/measurement pairs or a tuple `(x, y, params)`
+        where `params` is a dict of physics generator parameters to be loaded into the physics each iteration.
     :param None, torch.optim.lr_scheduler.LRScheduler scheduler: Torch scheduler for changing the learning rate across iterations.
     :param bool online_measurements: Generate the measurements in an online manner at each iteration by calling
         ``physics(x)``. This results in a wider range of measurements if the physics' parameters, such as
@@ -84,10 +92,11 @@ class Trainer:
         the measurements are loaded from the training dataset.
     :param None, deepinv.physics.generator.PhysicsGenerator physics_generator: Optional physics generator for generating
         the physics operators. If not None, the physics operators are randomly sampled at each iteration using the generator.
-        Should be used in conjunction with ``online_measurements=True``. Also see ``loop_physics_generator``.
-    :param bool loop_physics_generator: if True, resets the physics generator back to its initial state at the beginning of each epoch,
+        Should be used in conjunction with ``online_measurements=True``, no effect when ``online_measurements=False``. Also see ``loop_random_online_physics``.
+    :param bool loop_random_online_physics: if True, resets the physics generator **and** noise model back to its initial state at the beginning of each epoch,
         so that the same measurements are generated each epoch. Requires `shuffle=False` in dataloaders. If False, generates new physics every epoch.
         Used in conjunction with ``physics_generator``.
+    :param bool global_optimizer_step: If ``True``, the optimizer step is performed once on all datasets. If ``False``, the optimizer step is performed on each dataset separately.
     :param Metric, list[Metric] metrics: Metric or list of metrics used for evaluating the model.
         They should have ``reduction=None`` as we perform the averaging using :class:`deepinv.utils.AverageMeter` to deal with uneven batch sizes.
         :ref:`See the libraries' evaluation metrics <metric>`.
@@ -133,7 +142,8 @@ class Trainer:
     scheduler: torch.optim.lr_scheduler.LRScheduler = None
     online_measurements: bool = False
     physics_generator: Union[PhysicsGenerator, List[PhysicsGenerator]] = None
-    loop_physics_generator: bool = False
+    loop_random_online_physics: bool = False
+    global_optimizer_step: bool = True
     metrics: Union[Metric, List[Metric]] = PSNR()
     device: Union[str, torch.device] = "cuda" if torch.cuda.is_available() else "cpu"
     ckpt_pretrained: Union[str, None] = None
@@ -201,7 +211,7 @@ class Trainer:
         elif (
             self.physics_generator is not None
             and self.online_measurements
-            and self.loop_physics_generator
+            and self.loop_random_online_physics
         ):
             warnings.warn(
                 "Generated measurements repeat each epoch. Ensure that dataloader is not shuffling."
@@ -215,13 +225,13 @@ class Trainer:
             if wandb.run is None:
                 wandb.init(**self.wandb_setup)
 
-        if not isinstance(self.losses, list) or isinstance(self.losses, tuple):
+        if not isinstance(self.losses, (list, tuple)):
             self.losses = [self.losses]
 
         for l in self.losses:
             self.model = l.adapt_model(self.model)
 
-        if not isinstance(self.metrics, list) or isinstance(self.metrics, tuple):
+        if not isinstance(self.metrics, (list, tuple)):
             self.metrics = [self.metrics]
 
         # losses
@@ -280,15 +290,18 @@ class Trainer:
             self.loss_history = []
         self.save_folder_im = None
 
-        self.load_model()
+        _ = self.load_model()
 
-    def load_model(self, ckpt_pretrained: str = None):
+    def load_model(self, ckpt_pretrained: str = None) -> dict:
         """Load model from checkpoint.
 
-        :param str ckpt_pretrained: checkpoint filename. If `None`, use checkpoint passed to class. If not `None`, override checkpoint passed to class.
+        :param str ckpt_pretrained: checkpoint filename. If `None`, use checkpoint passed to class init.
+            If not `None`, override checkpoint passed to class.
+        :return: if checkpoint loaded, return checkpoint dict, else return ``None``
         """
         if ckpt_pretrained is None and self.ckpt_pretrained is not None:
             ckpt_pretrained = self.ckpt_pretrained
+            self.ckpt_pretrained = None
 
         if ckpt_pretrained is not None:
             checkpoint = torch.load(
@@ -302,6 +315,7 @@ class Trainer:
                 self.wandb_setup["resume"] = "allow"
             if "epoch" in checkpoint:
                 self.epoch_start = checkpoint["epoch"]
+            return checkpoint
 
     def log_metrics_wandb(self, logs: dict, step: int, train: bool = True):
         r"""
@@ -371,8 +385,8 @@ class Trainer:
 
         if self.physics_generator is not None:
             params = self.physics_generator[g].step(batch_size=x.size(0))
-            # Update parameters both via update_parameters and, if implemented in physics, via forward pass
-            physics.update_parameters(**params)
+            # Update parameters both via update and, if implemented in physics, via forward pass
+            physics.update(**params)
             y = physics(x, **params)
         else:
             y = physics(x)
@@ -417,7 +431,7 @@ class Trainer:
 
         if params is not None:
             params = {k: p.to(self.device) for k, p in params.items()}
-            physics.update_parameters(**params)
+            physics.update(**params)
 
         return x, y, physics
 
@@ -471,7 +485,9 @@ class Trainer:
 
         return x_net
 
-    def compute_loss(self, physics, x, y, train=True, epoch: int = None):
+    def compute_loss(
+        self, physics, x, y, train=True, epoch: int = None, backward=False
+    ):
         r"""
         Compute the loss and perform the backward pass.
 
@@ -482,12 +498,13 @@ class Trainer:
         :param torch.Tensor y: Measurement.
         :param bool train: If ``True``, the model is trained, otherwise it is evaluated.
         :param int epoch: current epoch.
+        :param bool backward: Whether to perform backward when computing the loss.
         :returns: (tuple) The network reconstruction x_net (for plotting and computing metrics) and
             the logs (for printing the training progress).
         """
         logs = {}
 
-        if train:
+        if train and backward:
             self.optimizer.zero_grad()
 
         # Evaluate reconstruction network
@@ -511,24 +528,27 @@ class Trainer:
                         self.logs_losses_train[k] if train else self.logs_losses_eval[k]
                     )
                     meters.update(loss.detach().cpu().numpy())
-                    cur_loss = meters.avg
-                    logs[l.__class__.__name__] = cur_loss
+                    logs[l.__class__.__name__] = meters.avg
 
-            meters = self.logs_total_loss_train if train else self.logs_total_loss_eval
-            meters.update(loss_total.item())
-            logs[f"TotalLoss"] = meters.avg
+                meters = (
+                    self.logs_total_loss_train if train else self.logs_total_loss_eval
+                )
+                meters.update(loss_total.item())
+                logs[f"TotalLoss"] = meters.avg
+        else:  # question: what do we want to do at test time?
+            loss_total = 0
 
         if train:
             loss_total.backward()  # Backward the total loss
 
-            norm = self.check_clip_grad()  # Optional gradient clipping
+            norm = self.check_clip_grad()
             if norm is not None:
                 logs["gradient_norm"] = self.check_grad_val.avg
 
-            # Optimizer step
-            self.optimizer.step()
+            if backward:
+                self.optimizer.step()  # Optimizer step
 
-        return x_net, logs
+        return loss_total, x_net, logs
 
     def compute_metrics(
         self, x, x_net, y, physics, logs, train=True, epoch: int = None
@@ -604,12 +624,20 @@ class Trainer:
             x_nl = y
         else:
             raise ValueError(
-                f"No learning reconstruction method {self.no_learning_method} not recognized"
+                f"No learning reconstruction method {self.no_learning_method} not recognized or physics does not implement it"
             )
 
         return x_nl
 
-    def step(self, epoch, progress_bar, train_ite=None, train=True, last_batch=False):
+    def step(
+        self,
+        epoch,
+        progress_bar,
+        train_ite=None,
+        train=True,
+        last_batch=False,
+        global_optimizer_step=True,
+    ):
         r"""
         Train/Eval a batch.
 
@@ -620,11 +648,15 @@ class Trainer:
         :param int train_ite: train iteration, only needed for logging if ``Trainer.log_train_batch=True``
         :param bool train: If ``True``, the model is trained, otherwise it is evaluated.
         :param bool last_batch: If ``True``, the last batch of the epoch is being processed.
+        :param bool global_optimizer_step: If ``True``, perform backward pass on all datasets before optimizer step.
         :returns: The current physics operator, the ground truth, the measurement, and the network reconstruction.
         """
+        if train and global_optimizer_step:
+            self.optimizer.zero_grad()  # Clear stored gradients
 
         # random permulation of the dataloaders
         G_perm = np.random.permutation(self.G)
+        loss = 0
 
         if self.log_train_batch and train:
             self.reset_metrics()
@@ -636,7 +668,15 @@ class Trainer:
             )
 
             # Compute loss and perform backprop
-            x_net, logs = self.compute_loss(physics_cur, x, y, train=train, epoch=epoch)
+            loss_cur, x_net, logs = self.compute_loss(
+                physics_cur,
+                x,
+                y,
+                train=train,
+                epoch=epoch,
+                backward=not global_optimizer_step,
+            )
+            loss += loss_cur
 
             # detach the network output for metrics and plotting
             x_net = x_net.detach()
@@ -651,6 +691,9 @@ class Trainer:
 
         if self.log_train_batch and train:
             self.log_metrics_wandb(logs, step=train_ite, train=train)
+
+        if train and global_optimizer_step:
+            self.optimizer.step()  # Optimizer step
 
         if last_batch:
             if self.verbose and not self.show_progress_bar:
@@ -707,7 +750,7 @@ class Trainer:
                 x_nl = None
 
             imgs, titles, grid_image, caption = prepare_images(
-                x, y, x_net, x_nl, rescale_mode=self.rescale_mode
+                x, x_net=x_net, x_nl=x_nl, rescale_mode=self.rescale_mode
             )
 
         if plot_images:
@@ -828,9 +871,13 @@ class Trainer:
                 [len(loader) - loader.drop_last for loader in self.train_dataloader]
             )
 
-            if self.loop_physics_generator and self.physics_generator is not None:
+            if self.loop_random_online_physics and self.physics_generator is not None:
                 for physics_generator in self.physics_generator:
                     physics_generator.reset_rng()
+
+                for physics in self.physics:
+                    if hasattr(physics.noise_model, "reset_rng"):
+                        physics.noise_model.reset_rng()
 
             self.model.train()
             for i in (
@@ -849,6 +896,7 @@ class Trainer:
                     train_ite=train_ite,
                     train=True,
                     last_batch=last_batch,
+                    global_optimizer_step=self.global_optimizer_step,
                 )
 
                 perform_eval = self.eval_dataloader and (
@@ -910,7 +958,13 @@ class Trainer:
 
         return self.model
 
-    def test(self, test_dataloader, save_path=None, compare_no_learning=True) -> dict:
+    def test(
+        self,
+        test_dataloader,
+        save_path: Union[str, Path] = None,
+        compare_no_learning: bool = True,
+        log_raw_metrics: bool = False,
+    ) -> dict:
         r"""
         Test the model, compute metrics and plot images.
 
@@ -918,6 +972,7 @@ class Trainer:
             a signal x or a tuple of (x, y) signal/measurement pairs.
         :param str save_path: Directory in which to save the plotted images.
         :param bool compare_no_learning: If ``True``, the linear reconstruction is compared to the network reconstruction.
+        :param bool log_raw_metrics: if `True`, also return non-aggregated metrics as a list.
         :returns: dict of metrics results with means and stds.
         """
         self.compare_no_learning = compare_no_learning
@@ -959,6 +1014,8 @@ class Trainer:
                 name = self.metrics[k].__class__.__name__ + " no learning"
                 out[name] = self.logs_metrics_linear[k].avg
                 out[name + "_std"] = self.logs_metrics_linear[k].std
+                if log_raw_metrics:
+                    out[name + "_vals"] = self.logs_metrics_linear[k].vals
                 if self.verbose:
                     print(
                         f"{name}: {self.logs_metrics_linear[k].avg:.3f} +- {self.logs_metrics_linear[k].std:.3f}"
@@ -967,6 +1024,8 @@ class Trainer:
             name = self.metrics[k].__class__.__name__
             out[name] = l.avg
             out[name + "_std"] = l.std
+            if log_raw_metrics:
+                out[name + "_vals"] = l.vals
             if self.verbose:
                 print(f"{name}: {l.avg:.3f} +- {l.std:.3f}")